+++ packages/i18n/Uppercase/Uppercase.js
@@ -12,6 +12,19 @@ import PropTypes from 'prop-types';
 
 import {toCapitalized, toUpperCase, toWordCase} from '../util';
 
+const isString = (content) => typeof content === 'string';
+
+const formatContent = (casing, content) => {
+	switch (casing) {
+		case 'word':
+			return toWordCase(content);
+		case 'sentence':
+			return toCapitalized(content);
+		case 'upper':
+			return toUpperCase(content);
+	}
+};
+
 /**
  * A higher-order component that is used to wrap an element to provide locale-aware uppercasing of
  * `children`, provided that `children` is a single string. Other values for `children` are
@@ -50,18 +63,10 @@ const Uppercase = hoc((config, Wrapped) => kind({
 
 	computed: {
 		children: ({casing, children}) => {
-			if (casing !== 'preserve' && React.Children.count(children) === 1) {
-				const content = React.Children.toArray(children)[0];
-				if (typeof content == 'string') {
-					switch (casing) {
-						case 'word':
-							return toWordCase(content);
-						case 'sentence':
-							return toCapitalized(content);
-						case 'upper':
-							return toUpperCase(content);
-					}
-				}
+			if (casing !== 'preserve' && React.Children.count(children)) {
+				return isString(children) ? formatContent(casing, children) : React.Children.map(children, (child) => {
+					return isString(child) ? formatContent(casing, child) : child;
+				});
 			}
 			return children;
 		}