+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -918,6 +918,24 @@ const VirtualListBase = VirtualListBaseFactory(JS);
 VirtualListBase.displayName = 'VirtualListBase';
 
 /**
+ * Activates the component for voice control.
+ *
+ * @name data-webos-voice-focused
+ * @memberof moonstone/VirtualList.VirtualListBase.prototype
+ * @type {Boolean}
+ * @public
+ */
+
+/**
+ * The voice control group label.
+ *
+ * @name data-webos-voice-group-label
+ * @memberof moonstone/VirtualList.VirtualListBase.prototype
+ * @type {String}
+ * @public
+ */
+
+/**
  * A Moonstone-styled base component for [VirtualListNative]{@link moonstone/VirtualList.VirtualListNative} and
  * [VirtualGridListNative]{@link moonstone/VirtualList.VirtualGridListNative}.
  *