+++ packages/sampler/stories/qa-stories/ContextualPopupDecorator.js
@@ -6,8 +6,10 @@ import React from 'react';
 import {storiesOf} from '@storybook/react';
 
 import {select} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 
 const ContextualButton = ContextualPopupDecorator(Button);
+const Config = mergeComponentMetadata('ContextualButton', ContextualButton);
 ContextualButton.displayName = 'ContextualButton';
 
 const buttonMargin = () => ({margin: ri.unit(12, 'rem')});
@@ -68,9 +70,9 @@ storiesOf('ContextualPopupDecorator', module)
 		() => (
 			<div style={{textAlign: 'center', marginTop: ri.unit(99, 'rem')}}>
 				<ContextualPopupWithActivator
-					direction={select('direction', ['up', 'down', 'left', 'right'], ContextualButton, 'down')}
+					direction={select('direction', ['up', 'down', 'left', 'right'], Config, 'down')}
 					popupComponent={renderPopup}
-					spotlightRestrict={select('spotlightRestrict', ['none', 'self-first', 'self-only'], ContextualButton, 'self-only')}
+					spotlightRestrict={select('spotlightRestrict', ['none', 'self-first', 'self-only'], Config, 'self-only')}
 				>
 					Hello Contextual Button
 				</ContextualPopupWithActivator>