+++ packages/core/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 The following is a curated list of changes in the Enact core module, newest changes on the top.
 
-## [unreleased]
+## [2.0.0-rc.1] - 2018-07-09
 
 ### Removed
 
-`core/util.withContextFromProps` function
+- `core/util.withContextFromProps` function
 
 ## [2.0.0-beta.9] - 2018-07-02