+++ packages/i18n/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 The following is a curated list of changes in the Enact i18n module, newest changes on the top.
 
-## [unreleased]
+## [2.0.0-rc.1] - 2018-07-09
+
+### Fixed
+
+- `i18n/Uppercase` to apply the designated `casing` prop format to each child instead of only the first child
 
 ### Removed